KG-47: Выход из группового чата

+++ server/api/chat/removeMember.ts
@@ -11,7 +11,13 @@ interface Params {
 
 export default async function(request: Request<Params>, response: Response) {
     const { chatId, userId } = request.params;
-    await findChat(request.user, chatId);
+    const chat = await findChat(request.user, chatId);
+
+    if (chat.type === 'dialog') {
+        // TODO: Придумать сообщение об ошибке
+        throw new Error('Chat is dialog');
+    }
+
     await Members.destroy({ where: { userId, chatId } });
 
     response.notification(chatId, 'REMOVE_MEMBER', { userId });
